Move central logic to module __init__, register template tag separately.
